+++ run_end_to_end_tests.py
@@ -49,6 +49,7 @@ log.setLevel(logging.ERROR)
 BASE_DIR = os.path.dirname(os.path.abspath(__file__))
 os.chdir(BASE_DIR)
 controller = None
+test_hermetics = False
 
 # Flask was unable to autofind the root_path correctly after an os.chdir() from another directory
 # Dunno why,refer to https://stackoverflow.com/questions/35864584/error-no-such-file-or-directory-when-using-os-chdir-in-flask
@@ -169,7 +170,8 @@ def start():
                      configs['input_config'],
                      configs['pipeline_config'],
                      configs['bitrate_config'],
-                     check_deps=False)
+                     check_deps=False,
+                     use_hermetic=test_hermetics)
   except Exception as e:
     # If the controller throws an exception during startup, we want to call
     # stop() to shut down any external processes that have already been started.
@@ -276,7 +278,13 @@ def main():
                       help='Number of trials to run')
   parser.add_argument('--reporters', nargs='+',
                       help='Enables specified reporters in karma')
+  parser.add_argument('--test-hermetics',
+                      action='store_true',
+                      help='Runs the tests using binaries from `shaka-streamer-binaries`')
   args = parser.parse_args()
+
+  global test_hermetics
+  test_hermetics = args.test_hermetics
 
   # Do static type checking on the project first.
   type_check_result = mypy_api.run(['streamer/'])